pyproject.toml
[project]
name = "triz-ai"
version = "0.1.0"
description = "AI-Powered TRIZ Innovation Engine — combines TRIZ methodology with AI and real patent data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15",
    "litellm>=1.50",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "pdfplumber>=0.11",
    "sqlite-vec>=0.1",
    "pyyaml>=6.0",
]

[project.scripts]
triz-ai = "triz_ai.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/triz_ai"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "pre-commit>=4.0",
    "pytest-timeout>=2.4.0",
]

[tool.ruff]
target-version = "py311"
src = ["src"]
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
